+++ src/commands/arg_structs/info_args.rs
@@ -6,12 +6,12 @@ pub struct ModInfoArgs {
     
     #[arg(short, long, value_name = "MODID")]
     pub mod_id: ModID,
+   
+    /// Shows the description of the mod, Note: This can take a lot of space in the terminal
+    #[arg(short = 'd', long)]
+    pub show_description: bool,
     
-    #[arg(short, long)]
-    pub show_versions: bool,
-    
-    #[arg(short, long, requires = "show_versions")]
-    pub last: Option<u32>,
-    
-    
+    /// Shows the last NUM amount of versions. Use with 0 to see ALL versions.
+    #[arg(short = 'v', long, default_value = "3", value_name = "NUM")]
+    pub show_versions: usize,
 }